Added relative touchpad motion.
Now if the user puts their finger down, the new motion will be
anchored relative to where they intially put their finger.
Additionally, if the user's cursor moves into the edges of the
screen, the anchor will be changed so that as soon as they move
away the cursor will start to move again, even if they "moved
into" the edge of the screen.